[project @ gilberto.echeverria@laas.fr-20100317171939-aiarvvvxd07ca4x5] Added the furniture files to scene/props. Though the items in these files do not have the collision envelopes.
Added a modified version of the cog_indoor_scene, with the latest versions of the scripts for Jido, the camera and the script holder. Still not tested yet.

Added the new camera, including the semantic functionality. This is in the files:
components/sensors/cameras/camera_component-2.1.blend
components/sensors/cameras/SemCamera_Module.py

Changes to the Teleport module

Added scripts and movement to the Jido model, in file:
components/robots/jido/jidosimu-2.0.blend