Correct DTM altitude for buildings's 1st Floor
Buildings' first floor's altitude is now estimated from the DTM itself
to avoid flying buildings.
We use the altitude of the DTM vertex the closest in 2D to the first
vertex of the building contour. An offset of -2m is applied so the building
foundations start underground. Wall height = 12m. Roof height = 15m.