Add a new default scene for use with MORSE
The new scene should be setup to be loaded every time Blender (MORSE)
runs. It imports the python script Setup_Morse_Paths.py, which
should correctly setup the python path to be used during the whole
simulation.

This way, other components will not need to worry about setting the path
again.

The new scene includes a First Person Camera and the Scene_Script_Holder
object, though this last one should be changed to eliminate references
to yarp or other middlewares.