Add blender files for new architecture components
A robot (dala), a sensor (gyroscope) and an actuator (keyboard control)
These new files make use of the object oriented python modules.
They implement a different system to call the methods of the python
modules, where a generic Python module is called, and behind the scenes
it identifies the correct object and calls the action method of that
object.