Add new python script files for objects
In the new architecture, all python files will reside inside the
$ORS_ROOT/scripts directory.

Morse_Main.py is the file called from the Scene_Script_Holder to do the
main part of controlling the simulation. It has been simplified from the
previous version (Scene_Module.py)

Morse_Object.py has the base class for all component objects in MORSE,
which can be robots, sensors or actuators. It contains methods and
variables that should be common to all components, such as the init
method, destructor method and an action method that does the bulk of
every component's job.

Morse_Calling.py is the wrapper for calling the methods of individual
objects. It is called from the Blender Logic Bricks, for any of the
components in the scene. It will then retrieve the object instance of
the component, and call its respective 'action' function.