Add the component-middleware binding to the .blend
A new text window in Blender contains the pyton script with a
dictionary. The dictionary lists all components and the middleware that
will be associated with each.
NOTE: The format for this dictionary is not yet fully defined.

The python file will be imported by the middleware classes that will
then add the hooks to the correct components.