Change to the selection of middleware by component
The binding is done by the init script of Morse_Main, instead of being
done by the Yarp module. This way all middlewares listed will be linked
with the components at the same time.

TODO: Will have to implement an abstract class for middlewares that
defines a common interface they should all have.