Change the configuration of components in blender
Change component_config.py file. Now the dictionary with the binding of
middlewares and components does not need the 'in/out' parameter.
The decision on how to configure the component is taken based on the
name of the function provided. Also, this is now done inside each
middleware, instead of the main.py module.