Change path in Blender files
Use the full morse path to class files in the Blender properties.
This will require that the ORS_ROOT variable points to the directory
that contains 'morse' with the python scripts.
For instance:
ORS_ROOT = $HOME/openrobots/lib/python2.6/site-packages

With this change, there should be no clash between the source and the
installed scripts