Add check of the success of opening posters
The succes of creating or finding a poster is returned as another value,
which Python will use to determine what to do next.
This is because otherwise the pointer of type POSTER_ID is not
understandable by Python, and its contents can not be used to make a
decision insido Python.