Change modifiers list in component_config.py
The list of modifiers included in the Blender files has changed.
The python file that lists which modifiers are applied to each sensor
has changed.
Now the dictionary contains elements indexed by the name of the
sensor/actuator, then a list of modifiers. Each modifier is specified as
another nested list, with the name of the object, and the function to be
added to the sensor.

NOTE: Currently, only the ROSACE scenario file complies with this change.