Add another parameter to 'init_extra_module'
This method is used to add functions to middlewares, beyond the most
basic data serialisation.
The new parameter for the function is the name of the method to be added
to a sensor or actuator. The name can be used in the function to
determine whether it should input or output data, and act accordingly.