Factorize the various CMakeLists.txt to build pocolibs
Now, you only need to modify a few variables in pocolibs/CMakeLists.txt
to add some new entries to the build.
It fixes some bugs too, because previously, we don't link with python
libs which can cause problems (particularly with python 3.1)
Another feature of this new CMakeLists.txt is to automatically generate
configuration variables to only build parts of pocolibs support.
You need to activate the variable
	BUILD_POCOLIBS_<MODULE>_SUPPORT
to activate the support for this module.

Most of the code comes from Nizar, thanks a lot to him :)