Change the arguments to add extra mw modules
The arguments passed to the functions that add external functionality to
middlewares have changed. Now the methods '_add_method' and
'init_extra_module' expect an array, which will have the list of
arguments given in 'component_config.py'.
This way we'll have acces to additional parameters for the middlewares.
The first such parameter that can be used is an optional name for the
pocolibs posters. Example (in component_config.py):

component_mw = {
		"Motion_Controller": ["Pocolibs", "read_genpos",
		"morse/middleware/pocolibs/controllers/genpos", "p3dSpeedRef"],
				}