Changes the kind of poster we export through pom interface
Previously, we export a POM_POS, similarly to the output of pom. Now, we
export a POM_ME_POS, which is an input of classic pom. In this way, we
can use the real pom module.

For component which export such a POM_ME_POS, we need to get two
parameters associated to the component : reference_frame (which is the
name name of the reference frame poster), and confidence ( the
confidence we have in the captor)