Reworked the Python path detection - no need to add anything to PYTHONPATH anymore
- The morse launch script checks for each path in the PYTHONPATH if 'morse/blender/main.py' exists.
If nothing is found, returns an error to the user, explaining what to do (...add the prefix to PYTHONPATH).
- Then, in Blender, the 'setup_path' script executed at MORSE startup now looks in the same way for the path where
MORSE python libraries are available, and dynamically add it to sys.path.

So if MORSE is installed in a 'normal' place (ie, a prefix already known by Python), it works out of the box.

While here:
 - fixed in default_scene the reference to setup_path (it was still looking for check_path)
 - in default_scene, moved all object above the ground (why were they below??)
 - improved the logging stuff in the morse launch script
 - fixed in grande_salle_kuka_2.55 the setup_path script
 - removed in the same scene the Yarp_Empty, because it would emit an error if MORSE was compiled
   without YARP. We need to find a way to properly handle case where some middleware are not available