Add a script to generate the scan arc of the SICK
Currently the SICK sensor uses the vertices in an arc mesh to determine
the direction of the rays to cast. The new script uses parameters on the
Sick Empty to generate a new arc.
However, this has to be done before running the Game Engine, and then
the arc has to be set in position.