Correct the length of the kinematic chain
It was being automatically computed as the number of children object,
but when something else is added at the end of the arm (such as a
hand) it didn't work anymore.
Now the length is considered from the array that contains the direction
in which the segments turn. This array is fixed size