Add component to heal near victims upon command
This component is a request of the Rosace project.
It consists of a Blender 'radar' object that will detect when a victim
is in fromt of the robot. If it receives a command, it will interact
with the victim and change its status.
The 'local_data' dictionary has only one item called 'heal'. It's normal
value is 0, and when set to 1, it will heal any nearby victims.