Rework how pocolibs middleware handle actuator process
Before, poster must exists when we initialize the game engine, so it
requires a strict ordering between start module, start simulator, init
module....
Now, we just register the poster to read from, and then, as long as we
don't find it, we look for it at each game engine turn. In this way, we
can start the simulator, even if modules are not here.

Common code is stored in PosterHandler.h and used by the different
specific poster. In the long run, we may need to build a library, and
link each module to it (here everything is in the header).

While here, remove crazy dependencies from lwr.