Correct the rotation of platine relative to robot
The rotation of the platine is now correctly handled even when the robot
rotates as well.
Also corrected the problem with modifying pan and tilt simultaneously.
For this correction, the platine object now also has two meshes that
rotate, just as the real unit has two separate segments. In the Blender
file, these are called 'PanBase' and 'TiltBase'. They must be included
in any scene using the platine.
The cameras on the robot must be parented to 'TiltBase' for them to
follow the movement of the platine.

Also modified the sample file:
examples/morse/scenarii/grande_salle-2.5-kuka-poster.blend

Conflicts:

	examples/morse/scenarii/grande_salle-2.5-kuka-poster.blend