Services can now be declared inside classes.
Any method defined in a class INHERITING FROM MorseObjectClass can now use the
'@service' decorator. It is registered with the name of the component instance as namespace.

Free functions declared as services must now set the 'component' parameter of the @service
decorator to set their namespace. For instance:

@service(component='simulation')
def restart_simulation():
   ...

While here, move the request managers initialization in main.py before
the initialization of the components.