Activate walk animations when controlling human
There is a new method in human_control.py that identifies if the human
is in motion, even when attached to a motion controller.
Also changed the phycis properties of the human in default_human.blend,
since the motion detection requires that the human be a dynamic object,
and that all bounding boxes for the body parts have the Compound flag for
dynamics.