Add Velodyne sensor and related pocolibs files
This is the first version of the Velodyne. It is based on how the
SICK is simulated. It makes use of an arc object to determine the
direction of the rays that will be cast to detect objects.
For the moment, the rotation of the velodyne is set directly
using the Logic Bricks.

The pocolibs middleware create a poster with the data, but have not
been fully tested yet.