Implement the 'goto' service for requests
Also change the control from position to velocity for robots.
This requires the main robot object to be either 'Dynamic' or 'Rigid
body'
Currently it is possible to change between the two types by uncommenting
the desired type within the script. However, this is only a temporary
solution.
To revert to position control, edit these lines as follows:
Change:
    self._type = 'Velocity'
    #self._type = 'Position'
for:
    #self._type = 'Velocity'
    self._type = 'Position'