Corrections to the service management
Make the configuration of the modules and services work the same way as
for the middlewares:
The script 'component_config.py' within the .blend file should have
another dictionary to specify which components use what service manager.
The dictionary must be called 'component_service'. Items in the
dictionary use the name of the component as the key, and the value is a
list, that contains the class name of the request manager. For example:

component_service = {
        "Motion_Controller": ["YarpRequestManager"],
            }