Multi node version of MORSE.
The changes in this commit will allow multiple instances of MORSE to
communicate the location of the robots in the scene.
There are several steps to setup a multi node simulation:
- Create 2 or more copies of the same Blender scenario, with the robots
that will be simulated on different nodes.
- In each blender file, change the 'Robot_Tag' property of the robots to
'External_Robot_Tag' for those robots that will be simulated in another
node. This should be done for every robot in each copy of the scenario.
- Configure the middleware bindings for the sensors on each copy of the
scenario. Remember there should be different names for the sensors and
actuators that will be active in each simulation node.
- Add to the Blender file a new text file, called: 'scene_config.py',
    with this content:
        node_config = {
            "node_name": "First node",
            #"server_address": "localhost",
            "server_address": "samus.laas.fr",
            "server_port": 65000
        }
- Change the 'node_name' in each file, as well as the 'server_address'.

- Start the Simulation Manager program. Currently there is a test program
in Python.

- Open the copies of the Blender scenario in different instances of
Blender/MORSE. The instances can be runnin on the same or different
computers.

- Start the simulation on all nodes. The positions of robots changed in
one node should be reflected in others.