Add empty robot as parent of standalone sensors
All sensors in a MORSE scene must be parented to a robot. Users can now
include this robot in a scene and parent standalone sensors to it.
The robot has no physical representation and can not move. It is only
represented by and Empty. The name of the robot is 'Environment'