Initial support for Pocolibs requests
This commit introduces a new request manager for the Pocolibs
middleware.

It do so by offering a custom implementation of the TCLserv protocol.
Hence, it's not /pre se/ a cslib clients, and applications relying on
the cslib library to directly send requests won't be able to use this
implementation.

Commands currently recognized:
 - HELLO
 - REPLYTO
 - RQST
 - BYE
 - LM (-> NOP)
 - ACK

Some raw TCL calls are also implemented (because they are used by the
pypoco Python bindings for TCLserv):
 - 'cs::lsmbox' (to list available modules)
 - 'info commands [module]::*Send' (to list available requests for a
   specific module)
 - cs::init (-> NOP)
 - exec (-> NOP)
 - module::connect (-> NOP)

Both synchronous and asynchronous request calls are supported.

TODO:
 - REPLYTO to another socket
 - Genom semantic (interrupted requests, etc)
 - infrastructure to use fake Genom modules (for name mapping)
 - and of course ...add services to components!