Change the name of the logger for error reporting
There is a conflict between the logger interface used when constructing
a simulation scene with the Builder API, and the logger used during the
simulation. This causes that after creating a scenario with Builder, it
will run properly only the first time the GE is launched. After presing
ESC to exit the simulation, presing P again will cause lots of error
messages indicating the logger could not be found.

This is only a problem when the scene is created using 'morse exec'.
If we save that scene as a .blend file, it can then be opened with morse
and the simulation run as many times as necessary without causing
problems.