Add simple human as a robot based on Rosace victim
It can be controlled by Actuators, such as waypoint or keyboard.
By default, it is set to be 'Injured' and will have a red color. The
color can be set to be more natural, by setting the 'Injured' property
to false, either in the Logic Bricks, or in the Builder script:
    target1 = Robot('victim')
    target1.properties(Injured = False)