Correct the rotation direction of the joints
This was tested against the model of the arm used in Move3D at LAAS.

Also added a cube to the base of the arm, to help determine the correct
orientation of the arm as a whole. This should match the orientation of
the base in the real arm.
In the case of the Jido robot, the box in the base is looking down. To
get the same orientation when using the Builder API, use these
transformations:
    kuka_arm = Actuator('kuka_lwr')
    kuka_arm.translate(x=0.1850, y=0.2000, z=0.9070)
    kuka_arm.rotate(x=1.5708, y=1.5708)
    Jido.append(kuka_arm)