Catch the mysterious errors in the logger
This modification will detect when the logger is being called and there
is no StreamHandler. In such a case it ignores the call, and will not
print unnecessary error messages.

However, this does not address the real problem, which is the fact that
a second logger is remaining hanging after creating a scene with the
Builder API. Yet no solution has been found to catch this problem at the
origin.