[RFC] Added unittesting for ROS services
This commit is a request for comments of a unittesting approach for
MORSE component.

The file morse/testing/middlewares/ros/testsuite.py contains:
 - a mini-scenario using the Builder API (try 'morse exec testsuite.py'
   to see),
 - an example of 2 unit-tests (for ROS services).

When run (./testsuite.py), the script starts Blender and calls itself
within Blender to construct the scenario, starts another process ('roscore'
in this example), and launches 2 tests (which won't pass right now: they
require changes in actuators/waypoint.py and services/supervision_services.py)

Open issues:
 - currently, the script does not kill the started external processes (blender,
roscore)
 - we need a way to know when morse is read (initialization done).

This commit also remove the 'internal ROS node' which is not necessary
anymore for the ros_request_manager. A new ROS node is created in
testing/ (obviously for testing purposes).