[testing] Set longer delays for the action server to settle
While here:
- re-enabled the test 'no action server'

NOTE REGARDING TESTING OF ROS ACTIONS
-------------------------------------

Currently, ROS' actionlib only support Python2.6.
Thus, the RosActionTest must be started with Python2.6.

This lead to a PYTHONPATH conflict for yaml (yaml is required in
both version 2.x and 3.x).

The current workaround is to manually prepend the path to yaml-2.x to
sys.path in the unit-test itself.

The real fix would be to add support for Py3 to actionlib.