Corrections to Segway robot
Included a commented part of the code to make the robot follow the
intended trayectory when controlled by the v_omega_differential_drive
actuator.
By modifying the variable _trackWidth, which should have the distance
between left and right wheels, we can get the expected trayectory from a
given v, omega. With this, the unit tests should pass.