Implement hack to correct the slip fo the robot
This is done by changing the value of the "distance between wheels" when
computing the wheel speeds. By doing this, it is possible to reduce the
slip of the wheels, though the exac values to use vary with the kind of
maneuver intended.
This parameter can be controlled from the Builder API. It is called
'FixTurningSpeed'