Add an Ignore property to not consider obstacles
The game property Ignore can be used to indicate a list of property
names, separated by commas.
If an object has any of these property names, it will not be considered
as an obstacle by the actuator, and will not force the robot to change
direction.
Useful for the victims in the ROSACE scenario, since the robot must
approach them directly to help them.