Implement the new object renaming policy
This commit introduce the automatic renaming of components,
based on the name of the variable used in the Builder script
and the place of the component in the hierarchy.

This should provide more natural naming behaviour (the components
become accessible with the same name as they were referenced in the
script) and prevent most of the name collision (that used to introduced
suffixes like '.001 etc)

Canvevat: only components visible from the top-level script, or
components appended to components visible from the scripts are renamed.