[sensor] Add depth_camera sensor
This sensor uses the z-buffer information from the VideoTexture module
in Blender to recover distances from the camera.
This will be the base for simulating laser range sensors such as SICK
and Velodyne.

NOTE: Using this sensor requires a patch for Blender, to be able to
recover the z-buffer data