[sensors] Update the depth camera
Decode the depth information comming from the patched Blender.
Data should be stored as a single unsigned int of 4 bytes instead of the usual
RGBA of the VideoTexture module.
The sensor will convert this value back to a float and do the conversions to
get a 3D point for each pixel.

This requires a patch to Blender, which will be added in the next commit