[pymorse] Complete rewrite of the pymorse bindings
This commit closes issue #216. Please refer to the issue for
discussion and a summary of the API changes.

Refer also to the new pymorse documentation for details.

In short:
- for each robot in the simulation, we dynamically create one object in the 'pymorse.Morse' instance
- this object itself contains one sub-object per component
- these sub-objects expose a datastream API, either read-only (for
  sensors): 'get', 'last', 'subscribe'; or write only (for actuators):
  'publish'.
- if exposed by the component in the simulator, RPC services are also
  available, as 'futures'. As a consequence, pymorse NOW REQUIRES
  PYTHON3. Note that cancelling a running service is currently not
  implemented.

Two unit-tests have been added in bindings/pymorse/ that cover most of
the new features.

This commit also introduces a new service 'simulation.details'. This
service returns the tree of robots, components, services, middlewares in
the current simulation