[sockets] Fix processing of concomittent socket requests
Because we were using 'readline' only once on sockets ready to
be read, we could miss requests when several came in a row.

The Python 'readline' API is however blocking.

Rewritten this part to use low-level 'recv' and manually
splitting the buffer into independant requests.

This assume the whole requests arrive as a single TCP package.