[pymorse] Support new component naming scheme
With the auto-renaming patch in, MORSE components now have a consistent naming scheme with
the full cmpt hierarchy.

This commit add support for this component hierarchy in pymorse.

For example, if you put a gripper at the tip of an arm, previously you had:

>>> robot.gripper.grab()

Now you have:

>>> robot.arm.gripper.grab()

This should prevent some name collisions.

Updated gripper_testing to show the new API.