[armature] Add a 'joint trajectory' interface
The new service (armature.trajectory(traj)) allow for execution
of trajectories on an armature: user provide a set of waypoints and
timestamps that the actuator try to follow.

The trajectory structure is similar to ROS' JointTrajectory.

This first version does not support velocities and accelerations, only
positions.

Refer to armature_testing.py for usage examples.