Introduce 'abstraction levels' for component
This commit introduces a new way to define the semantics of a component in
terms of abstraction levels or 'realism' levels.

By adding 'add_level' statements in a component declaration, one may
create new levels that correspond to a specific subset of exported
data fields and/or a custom component implementation.