[pocolibs] Fix pose computation for images
If there are some different frequencies on stereo and video_camera, it
is possible that the robot position when the shots are taken is not the
same than the one when we output data. So store properly the robot
position in camera_video, and use it to compute the transformation for
viam output.

While here, fix an odd rotation paramter.