[pr2] Updates to PR2 blender model
- Adjusted torso armature to have correct '0'
- Reparented torso_lift children to torso_armature

This would previously prevent the head and arms to move along with the torso.
It actually looks like a bug in the BGE (or at least, an inconsistency).

This change also required clearing the parent inverse matrix of head and arms, and
updating accordingly transformations of these parts. This have introduce small
inaccuracies in bodies placement compared to original URDF.