[doc] Simulation creation tutorial, chapt. 0->4
This large tutorial covers the process of creating a complex
simulation in MORSE, with custom robot and actuators. It is
based on the EPFL's Ranger project.

It covers Blender modelling as well as Python programming and client
side testing with pymorse.

This commit provides chapters 0 to 4:

Chapter 1: The Ranger robot
Chapter 2: Making a robot mesh
Chapter 3: Our first robot controller
Chapter 4: Creating a new actuator to move the eyes