[builder] Try to use default interface whenever possible
When setting a default interface to a robot, and adding a component
that is not declared in builder/data.py, it fails.

This commit tries to use, if it exists, the default method of the datastream mw.

Typical scenario: you develop a new component out-of-tree. You want to
test it/debug it with the socket interface. It may work out-of-the-box
(we still display a warning, though).