[bin] Export the whole sys.path in PYTHONPATH to Blender
The sys.path is not initialized in the same way with the python3
interpreter, and the inner python3 embedded in blender. So export fully
the sys.path through PYTHONPATH so the user get the expected environment
in blender context.

It was done before, but lost in 987162b787. Reported by Felix Ruess.