[testing] Support testing for invalid scenes
This commit introduces a new class of tests: MorseBuilderFailureTestcase. These
tests pass if MORSE *properly fails* to load the scene (ie, it tests that the
GameEngine is not started).

The commit also include 2 tests that currently do not pass (cf bug #374).

While here:
 - improve management of PIDs (you do not need to get it from MORSE log,
   Popen.subprocess give it back to us). This allow to kill MORSE even
   if it did not reach the initialization stage.

TODO: I've no example of invalid scripts that are correctly handled by
MORSE, so I could not test that MorseBuilderFailureTestcase actually
pass when MORSE dies before starting the BGE.