[sensor] cleanup Kinect still WIP
The idea is to use ROS' depth_image_proc package to mix depth and rgb
images, instead of converting the OpenGL Depth (Z) buffer to an XYZ
point-cloud, we send a float32 depth image as the kinect does (it sends
uint16 instead, just have to skip `depth_image_proc/convert_metric`[1]).

[1] http://wiki.ros.org/depth_image_proc#depth_image_proc.2BAC8-convert_metric

As mentioned in #419, the launch-file is still a work in progress,
included here as a base for future work.

[#419] https://github.com/morse-simulator/morse/issues/419