[builder] Rework implementation of add_default_interface
The new implementation only adds the default service and datastream
entry if the sensor | actuator does not get already some service or
datastream configured for this specific middleware. Moreover, it now
handles properly the sensors and actuators added after the call of
add_default_interface.

It should fix #399 (but people interested by the features must test to
see if it matches their expected semantics).