[robots/grasper] modified grasp service for grasper robot.
The service request grasp_ was modified into grasp(True,"object_name") in grasper.
It allow a robot to grasp a defined object and attach it to the arm.
If no name is specified it will grasp the 1st object detected by the
Near sensor.
If True is given as 1st parameter, it will grasp the object.
If False, it will release it from robot's hand.

Also the RobotGrasper class has been renamed into GraspingRobot.

Change grasp service parameter.