[builder] Removed builder.creator.RobotCreator
This class simply created an empty robot, but was broken because
the resulting component did not inherit of builder.Robot (one of the
consequences is that a RobotCreator instance is not considered for
automatic renaming).

Better to remove it completely.
Now, when creating a robot *without a Blender model*, an Empty is added.
This makes the FakeRobot (the only user of RobotCreator) a real robot, and
should fix bugs (because, as a RobotCreator instance, it was not considered
for automatic renaming, and thus its sensors/actuators would not be
accessible from pymorse for instance).