[template] More fix for default velocity_model
Fix the template blend file so it works properly with physics control.
It could have been corrected directly in the builder template file, but
it would make it harder to understand for new users.
In the template builder file, set the default controltype of keyboard to
Position, so it can work in parallel of VWMotion (see also d27a0741243)