[sensor/camera] Handle more cleverly relations between objs of different scenes
Contrarily to what we can expect, several objects can have the same name
in the same scene. So we need to track the relation between the objects
of the different scene in a smarter way. As said in the code comment,
I'm not completely sure it is totally correct, but it largely improve the
situation, and I can't see any regression.

Should fix #509