[camera] Add scene only if the camera is owner by an 'internal' robot
Fix #526

The semantic of bge.logic.addScene is crappy (and I'm generous). If the
scene does not exist, it still do something, which seems basically
duplicate the current scene, with all the logic. It is a problem in the
multinode case where only the scene related to the controlled robot are
created.

Still, it is definitively not a good fix, as it leaks multinode
abstraction in a component. The 'good' solution would be to 'not
configure' component owned by external robot (instead of configure, then
disable), but last time I tried that, it seems a lot more harder than
expected.