[sensors] Add a property ComputationMethod for accelero, imu, velocity
Instead of relying only on the fact that a robot has physics to choose
the way the velocity and the acceleration are computed, use a property
allowing the user to configure it. The default (without any
configuration from user) does not change.

In particular, it permits to make these sensors to work properly for
robot with dynamic, but using actuator with ControlType = 'Position'.