[doc] Reworked documentation of environments
Commit 6b8726391c1a662 introduced a gallery of available 3D
environment, but it was broken (issue #631) because since the
gallery was auto-generated by a sphinx extension, the URL of images
never appeared, and sphinx wouldn't copy the required images.

This commit actually generate (in generate_doc) one doc page per available
environment (in, all *.blend in data/environments) with a picture.

Hence, the pictures are listed, copied, and displayed in the gallery.

While here, moved 2 images of the sandbox and outdoors environment to
the correct location, and added a link the the gallery of environments
in the component library.

Fix #631