[actuator] Fix fallout from b70d5c1a9
Now, all the sensors / actuators from a scene are read a firs time
outside the blender context (as we need to read their class and parent
class from builder). Pulling symbol directly from blenderapi is not a
good idea as they are not refreshed when we force the reimport of
blenderapi (once in game). So, just import blenderapi and read symbols
from it.