[core] Improve correctness of time handling in BestEffortStrategy
The implementation was correct in nominal case, basically when you have
one physical / logical step for one rendering step. In other cases,
the logic of Blender is a bit strange, and it is hard to guess what is
the real timestep between two frames. Exporting it  through an API would
be the good solution, but D728 is always in discussion.

So observe it using an indirect way. Put a solid at a certain speed, and
measure what is its displacement between this two frames. Assuming the
physical engine is perfect, we can compute the delta t between two
frames.