[coordinates] Prepare Morse to handle a transformation between blender frame and ltp frame
In some situation, we need to have a transformation between blender
frame and ENU frame. Still, for some sensors, we need to be able to
compute in the ltp frame (for magnetic field, and so on). Prepare the
infrastructure to handle this transformation (no modifier are expected
in behaviour through).