Corrected/improved the English
I find this unclear (and it should be according not accordingly):

- **Rigid Body**: The object will behave accordingly to its shape.

Do you mean:

- **Rigid Body**: The object will behave according to its shape's
  properties.

It seems confusing to me that Blender has Sensors and Actuators and so
does morse but the morse ones are different. Pity the morse ones aren't
called, say BotSensor(s) and BotActuator(s), but I guess it is too late
now:-)

I'm surprised you haven't added an option to Blender to select a whole
robot since it is tricky, but perhaps that isn't possible.