[doc] Corrected/improved the English
For rqt_plot the x and y arguments are , separated but the z is space
separated which seems odd to me?

I changed 'paste the following code' to 'type the following code'
because if you paste a lump of code like this into the REPL it won't
work unless you do it line-by-line (plus one block for the function).