Obeject server changes
Deprecated can be used for older optix (5)
Updated can be used for optix 7 but will fail silently if cortex is not installed

There were issues with morse.builder import * trying to import cortex and failing
if it was not installed.

Now that I am writing the commit i am wondering if the fail silently is a bad idea.
The main reason was because a function definition used a cortex type that wasnt
getting included and would fail - but if that definition is changed then there is no need
to fail silently