Added a new range sensor
Unlike the altitude or FLS sensors, this range sensor makes no assumptions about the direction that it is ranging in and bases it's
position and orientation off a named blender object. This lets a user define the orientation in a .blend file and not in code.

Looks up the blender name for unique sensor names (Used in the MOOS middleware notifier)